fixed bugs

+++ src/main/java/bzu/edu/hotelManagmentAPI/service/RoomServiceImp/RoomServiceImp.java
@@ -5,11 +5,14 @@ import bzu.edu.hotelManagmentAPI.dto.RoomRequestDto;
 import bzu.edu.hotelManagmentAPI.dto.RoomResponseDto;
 import bzu.edu.hotelManagmentAPI.dto.RoomUpdateDto;
 import bzu.edu.hotelManagmentAPI.repository.RoomRepository;
+import bzu.edu.hotelManagmentAPI.security.SecurityUtils;
 import bzu.edu.hotelManagmentAPI.service.RoomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
+import org.springframework.stereotype.Service;
 
+@Service
 public class RoomServiceImp implements RoomService {
     private final RoomRepository roomRepository;
 
@@ -41,6 +44,7 @@ public class RoomServiceImp implements RoomService {
 
     @Override
     public void deleteRoom(Long id) {
+        SecurityUtils.checkIfAdminAuthority();
         roomRepository.deleteById(id);
     }
